[JIT] Add a python hook for a function to interpret JIT graphs.
This function allows to execute a graph exactly as it is, without going
through a graph executor which would run passes on the graph before
interpreting it. I found this feature extremely helpful when I worked on
a stress-testing script to shake out bugs from the TE fuser: I needed to
execute a very specific set of passes on a graph and nothing else, and
then execute exactly it.

ghstack-source-id: 77d6e33e5deb522e8ae2f6e645ee43852f2bd516
Pull Request resolved: https://github.com/pytorch/pytorch/pull/44493